cos/queue_ops.rs: colocate 49 MQFQ + V-min + flow-fair tests with production (#984 P3 Phase 5a)
Largest single phase. Move 49 tests from tx/mod.rs::tests to a new
mod tests at the bottom of cos/queue_ops.rs. Tests cover MQFQ (28),
V-min suspension state machine (17), and flow-fair plumbing (4).

Imports added in the new mod tests block: PROTO_TCP, tx_frame_capacity,
MmapArea, the production CoSQueueConfig/FastMap/FlowRrRing/Prepared*/Tx*
types, COS_MIN_BURST_BYTES, ExactCoSScratchBuild, the admission +
queue_service helpers the tests exercise, plus the cos_classify helpers
(cos_queue_accepts_prepared, demote_prepared_cos_queue_to_local).

Production-side widening: cos_queue_accepts_prepared and
demote_prepared_cos_queue_to_local pub(super) → pub(in crate::afxdp)
so cos/queue_ops::tests can reach them via the crate path. Justified:
the alternatives (move tests back to cos_classify::tests, or add cfg-test
wrappers) add more code surface than the visibility delta itself.

Test bodies that referenced super::super::worker rewritten to
crate::afxdp::worker (same module, explicit path required at the new
nesting depth).

Drops ~25 stale cfg-test imports from tx/mod.rs (Codex round-1 cleanup).
Added cos_flow_aware_buffer_limit back to the production import block —
tx/cos_classify.rs:739 reaches it via use super::*.

test_support.rs now imports build_cos_interface_runtime directly from
cos/builders since the cfg-test re-export through tx/mod.rs is gone.

Rebased onto current master after #1009 + #1011 merged.
cargo test --bins 865/0/2.
